Add search command
Search opens a new window, writes a prompt, and listens for insert
or delete events which it reflects in its prompt location and query
content. As the query is entered, an search is performed using
`L sym -p`, which prints the contents and locations of global symbol
search results -- debounced for 200ms. The output is parsed, scored,
and added to a heap. Every 100ms or after `L sym` exits, if the search
has not been canceled by an edit, the top 20 results are printed.
The results and locations in buffer are stored so that right click
events can be routed to plumber via executing `plumb` with the address
part of `L sym -p` output.