Place prompt up top
Placing the prompt at the top means that the cursor will move to the
bottom of the window when writing new results, and will have to be reset.
However, since the prompt moves around less, insert and delete events
don't race with result display writes and the position of inserts/deletes
is stable.